pyproject.toml
[project]
name = "sales_contact_finder_crew"
version = "0.1.0"
description = "Sales contact finder crew using crewAI"
authors = [{ name = "GibsonAI", email = "team@gibsonai.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.108.0,<1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0"
]

[project.scripts]
sales_contact_finder_crew = "sales_contact_finder_crew.main:run"
run_crew = "sales_contact_finder_crew.main:run"
train = "sales_contact_finder_crew.main:train"
replay = "sales_contact_finder_crew.main:replay"
test = "sales_contact_finder_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
